Adjust oauth entities to work consistently

+++ concrete/src/Entity/OAuth/AuthCodeRepository.php
@@ -3,6 +3,7 @@
 namespace Concrete\Core\Entity\OAuth;
 
 use Concrete\Core\Entity\Express\EntityRepository;
+use Doctrine\ORM\EntityManager;
 use Doctrine\ORM\EntityManagerInterface;
 use League\OAuth2\Server\Entities\AuthCodeEntityInterface;
 use League\OAuth2\Server\Repositories\AuthCodeRepositoryInterface;
@@ -28,8 +29,8 @@ class AuthCodeRepository extends EntityRepository implements AuthCodeRepositoryI
      */
     public function persistNewAuthCode(AuthCodeEntityInterface $authCodeEntity)
     {
-        $this->getEntityManager()->transactional(function(EntityManagerInterface $em) use ($authCodeEntity) {
-            $em->persist($authCodeEntity);
+        $this->getEntityManager()->transactional(function(EntityManagerInterface $entityManager) use ($authCodeEntity) {
+            $entityManager->persist($authCodeEntity);
         });
     }
 
@@ -41,9 +42,14 @@ class AuthCodeRepository extends EntityRepository implements AuthCodeRepositoryI
      */
     public function revokeAuthCode($codeId)
     {
-        $this->getEntityManager()->transactional(function(EntityManagerInterface $em) use ($codeId) {
-            $code = $em->find($codeId);
+        $code = $this->find($codeId);
 
+        if (!$code) {
+            throw new \InvalidArgumentException('Invalid auth token code');
+        }
+
+        $this->getEntityManager()->transactional(function(EntityManagerInterface $em) use ($code) {
+            $code = $em->merge($code);
             if ($code) {
                 $em->detach($code);
             }
@@ -59,6 +65,6 @@ class AuthCodeRepository extends EntityRepository implements AuthCodeRepositoryI
      */
     public function isAuthCodeRevoked($codeId)
     {
-        return (bool) $this->find($codeId);
+        return $this->find($codeId) === null;
     }
 }
